+++ src/main/java/sangmyung/chatprompt/task/repository/IoPairRepository.java
@@ -14,5 +14,5 @@ public interface IoPairRepository extends JpaRepository<IOPairs, Long> {
     List<IOPairs> findPairsByTaskId(@Param("taskId") Long taskId);
 
     @Query("select i from IOPairs i where i.task.id =:taskId and i.idx =:ioIndex")
-    Optional<IOPairs> findPairByIoIndex(@Param("taskId") Long taskId, @Param("ioIndex") Long ioIndex);
+    Optional<IOPairs> findPairByIoIndex(@Param("taskId") Long taskId, @Param("ioIndex") int ioIndex);
 }
